backend/requirements.txt
# Core dependencies
python-telegram-bot>=20.0
aiohttp>=3.8.1
Django>=4.2.0
psycopg2-binary>=2.9.3
redis>=4.3.4
celery>=5.2.7

# Security
cryptography>=37.0.0
pyOpenSSL>=22.0.0
requests>=2.28.1
aiohttp>=3.8.1
python-geoip>=1.2
maxminddb-geolite2>=2018.703
proxycheck>=0.2.0

# Monitoring and Analytics
prometheus_client>=0.14.1
statsd>=3.3.0
datadog>=0.44.0
sentry-sdk>=1.9.0

# Database
motor>=3.0.0
pymongo>=4.2.0
redis>=4.3.4

# Utilities
python-dateutil>=2.8.2
pytz>=2022.1
pycountry>=22.3.5
python-dotenv>=0.20.0
ujson>=5.4.0

# Testing
pytest>=7.1.2
pytest-asyncio>=0.18.3
pytest-cov>=3.0.0
pytest-mock>=3.8.2

# Documentation
Sphinx>=5.0.2
sphinx-rtd-theme>=1.0.0

# Development
black>=22.3.0
flake8>=4.0.1
mypy>=0.961
isort>=5.10.1

# Additional dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aioredis==2.0.1
segno==1.5.2
jdatetime>=5.2.0
python-telegram-bot==20.7
httpx==0.25.2
pytz==2023.3
pydantic==2.6.1
python-dateutil==2.8.2
jdatetime==4.1.1
persiantools==3.0.1
emoji==2.8.0
qrcode==7.4.2 